Add a "IsManaged" entry to getContainers
Also added a directory for containers.
This will be useful to save all the containers' compose files. If a compose file is inside this directory, it's considered "Managed", which allows Bayt to edit or delete the compose file, or create a new one.

This is as to not potentially break user configs and have a single controlled(ish) environment for containers' compose files, where the user also expects the files to be editable by Bayt and its clients.